+++ Sources/TMDB_Shared_Backend/Models/AccountInfoModel.swift
@@ -1,0 +1,22 @@
+struct AccountInfoModel: Codable {
+    let avatar: Avatar
+    let id: Int
+    let iso_639_1: String
+    let iso_3166_1: String
+    let name: String
+    let include_adult: Bool
+    let username: String
+    
+    struct Avatar: Codable {
+        let gravatar: Gravatar
+        let tmdb: TMDB
+        
+        struct Gravatar: Codable {
+            let hash: String
+        }
+        
+        struct TMDB: Codable {
+            let avatar_path: String?
+        }
+    }
+}
