+++ Sources/TMDB_Shared_Backend/TMDBAPIError.swift
@@ -1,0 +1,7 @@
+import Foundation
+public enum TMDBAPIError: Error {
+    case noResponse
+    case jsonDecodingError(error: Error)
+    case networkError(error: Error)
+    case unsupportedEndpoint
+}
